Adding motionrc
