+++ .motionrc.yml
@@ -1,0 +1,5 @@
+MOTION_REDIS_DB: 0
+MOTION_REDIS_HOST: localhost
+MOTION_REDIS_PASSWORD: null
+MOTION_REDIS_PORT: 6381
+MOTION_REDIS_SSL: false
